feat(vincent-rest-api): API Definitions fixes / enhancements - Extract schemas and registry entries into separate files with clearer dependencies --> no longer all contained in a single 2000 line file that has brittle ordering constraints due to 'no reference before defined' rules - Eliminated extraneous type exports that shouldn't have been exported or referenced externally - Ensured mutations to registry are done explicitly (not as part of loading the module) - Seperate files for each domain slice for both schemas and routes - Changed some names to eliminate `Def` for things that aren't actually definition docs - Fixed payload types for creating new appVersion and toolVersion (they were missing packageName and version fields)
