+++ packages/vincent-rest-api/src/index.ts
@@ -1,4 +1,4 @@
-import { registry as openAPIRegistry } from './lib/apiDefinition';
+import { registry as openAPIRegistry } from './lib/baseRegistry';
 import openApiJson from './generated/openapi.json';
 
 export { vincentApiClient } from './generated/vincentApiClient';
